+++ client/app/directives/tree_directive/TreeDirective.js
@@ -156,7 +156,6 @@ angular.module('wasdi.TreeDirective', [])
                                     if(data.node.original.band.bVisibleNow == true)
                                     {
                                         data.node.original.band.bVisibleNow = false;
-                                        //$('#jstree').jstree(true).set_icon(data.node.id, 'assets/icons/uncheck.png');
                                         $('#jstree').jstree(true).set_icon(data.node.id, 'assets/icons/uncheck_20x20.png');
                                         $scope.m_oController.removeBandImage(data.node.original.band);
                                     }
@@ -166,8 +165,6 @@ angular.module('wasdi.TreeDirective', [])
                                         //the tree icon is change when it receive the "publishband" message by rabbit or
                                         //when the band was pubblished (http request)
                                         // method: receivedPublishBandMessage()
-                                        //data.node.original.bVisibleNow = true;
-                                        //$('#jstree').jstree(true).set_icon(data.node.id, 'assets/icons/check.png');
                                         $scope.m_oController.openBandImage(data.node.original.band);
                                     }